Moved KeyManager to crypto package
